pyproject.toml
[project]
name = "devlog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "markdown>=3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
